Improve handling of the workspace directory
The workspace dir can be legitimately empty
Use a tmpdir for the workspace, as the tests can end up with multiple
instances using the same one which leads to race conditions
Create new workspace on server restart
Delete the project, implementing the log handler to populate and put in
ycmd's log

+++ ycmd/completers/java/java_completer.py
@@ -26,6 +26,7 @@ import logging
 import os
 import threading
 import glob
+import tempfile
 
 from shutil import rmtree
 from subprocess import PIPE
@@ -47,16 +48,6 @@ LANGUAGE_SERVER_HOME = os.path.join( os.path.dirname( __file__ ),
                                      'repository')
 
 PATH_TO_JAVA = utils.PathToFirstExistingExecutable( [ 'java' ] )
-
-# TODO: If there are multiple instances of ycmd running, they will _share_ this
-# path. I don't think (from memory) that eclipse actually supports that and
-# probably aborts
-WORKSPACE_PATH_BASE = os.path.join( os.path.dirname( __file__ ),
-                                    '..',
-                                    '..',
-                                    '..',
-                                    'third_party',
-                                    'eclipse.jdt.ls-workspace' )
 
 
 def ShouldEnableJavaCompleter():
@@ -115,14 +106,12 @@ class JavaCompleter( language_server_completer.LanguageServerCompleter ):
     # Used to ensure that starting/stopping of the server is synchronised
     self._server_state_mutex = threading.RLock()
 
+
     with self._server_state_mutex:
       self._server = None
       self._server_handle = None
       self._server_stderr = None
-      self._workspace_path = os.path.join(
-        os.path.abspath( WORKSPACE_PATH_BASE ),
-        str( os.getpid() ) )
-
+      self._workspace_path = None
       self._Reset()
 
       try :
@@ -191,15 +180,18 @@ class JavaCompleter( language_server_completer.LanguageServerCompleter ):
         utils.RemoveIfExists( self._server_stderr )
         self._server_stderr = None
 
+    if self._workspace_path:
+      try:
+        rmtree( self._workspace_path )
+      except OSError:
+        # We actually just ignore the error because there's really not much
+        # else we can do
+        _logger.exception( 'Failed to remove workspace path: {0}'.format(
+          self._workspace_path ) )
+
+    self._workspace_path = tempfile.mkdtemp()
     self._server_handle = None
     self._received_ready_message = threading.Event()
-
-    try:
-      rmtree( self._workspace_path )
-    except OSError:
-      # We actually just ignore the error because on startup it won't exist
-      _logger.exception( 'Failed to remove workspace path: {0}'.format(
-        self._workspace_path ) )
 
     self._server = None
